requirements.txt
facebook-scraper==0.2.57
python-dotenv==0.20.0
requests==2.27.1